+++ LefDefFilesReader/lib/containers/rect/Rect.h
@@ -5,15 +5,15 @@ public:
 	Point First;
 	Point Second;
 
-	Rect(double firstX = 0.0, double firstY = 0.0, double secondX = 0.0, double secondY = 0.0);
+	Rect(int firstX = 0.0, int firstY = 0.0, int secondX = 0.0, int secondY = 0.0);
 	Rect(Point& first, Point& second);
 	Rect(const Rect& value);
 
 	void SetFirst(const Point& value);
-	void SetFirst(const double first, const double second);
+	void SetFirst(const int first, const int second);
 
 	void SetSecond(const Point& value);
-	void SetSecond(const double first, const double second);
+	void SetSecond(const int first, const int second);
 
 	Point GetFirst() const;
 	Point GetSecond() const;
